Convert disabled VT's DnD tests to interactive UI tests
Move VerticalTabStripDragAndDropBrowserTest's three disabled tests
(DragTabToReorder, DragTabToDetach, DragURL) from brave_browser_tests
to the new brave_interactive_ui_tests suite, and delete the originals.
