+++ browser/ui/views/frame/vertical_tabs/BUILD.gn
@@ -3,9 +3,6 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import("//build/config/ozone.gni")
-import("//build/config/ui.gni")
-
 source_set("vertical_tabs") {
   sources = [
     "vertical_tab_strip_region_view.cc",
@@ -43,6 +40,26 @@ source_set("vertical_tabs") {
   ]
 }
 
+source_set("interactive_ui_tests") {
+  testonly = true
+  defines = [ "HAS_OUT_OF_PROC_TEST_RUNNER" ]
+
+  sources = [ "vertical_tab_strip_interactive_ui_tests.cc" ]
+
+  deps = [
+    ":vertical_tabs",
+    "//brave/browser/ui/views/frame",
+    "//chrome/browser",
+    "//chrome/browser/ui",
+    "//chrome/browser/ui/views/location_bar",
+    "//chrome/test:test_support_ui",
+    "//content/test:test_support",
+    "//testing/gtest",
+    "//ui/base:test_support",
+    "//ui/views",
+  ]
+}
+
 source_set("browser_tests") {
   testonly = true
   defines = [ "HAS_OUT_OF_PROC_TEST_RUNNER" ]
@@ -59,7 +76,6 @@ source_set("browser_tests") {
     "//chrome/browser",
     "//chrome/browser/tab_group_sync:factories",
     "//chrome/browser/ui",
-    "//chrome/browser/ui/views/location_bar",
     "//chrome/test:test_support_ui",
     "//content/test:test_support",
     "//testing/gtest",
@@ -69,15 +85,4 @@ source_set("browser_tests") {
     "//ui/views",
     "//ui/views:test_support",
   ]
-
-  if (use_aura) {
-    deps += [ "//ui/aura:test_support" ]
-  }
-
-  if (use_ozone) {
-    deps += [
-      "//ui/ozone",
-      "//ui/platform_window/common",
-    ]
-  }
 }
